pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Chikubi-Kun <raiseyourtail@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sanic = "^23.6.0"
sanic-jwt = "^1.8.0"
strawberry-graphql = "^0.213.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
